Word claims about key commitment more carefully
In the meeting of 2022-06-02,
Daniel Bleichenbacher advised to make the claims about key commitment
more carefully.
There is no research that anyone in the meeting at the time was aware of
that showed AEGIS to be key-committing.
Key commitment was not part of the AEGIS design criteria at the time,
if perhaps only because the paper on partitioning oracle attacks hadn't
even been published at the time.

Therefore, invert claim from AEGIS being key-committing to it not having
been shown that it is not key-committing.